Explicit casts in atomic return types

+++ include/crstl/atomic.h
@@ -23,132 +23,132 @@ crstl_module_export namespace crstl
 
 	inline int8_t atomic_store(int8_t volatile* target, int8_t value) crstl_noexcept
 	{
-		return crstl_atomic_store8(target, value);
+		return (int8_t)crstl_atomic_store8(target, value);
 	}
 
 	inline int16_t atomic_store(int16_t volatile* target, int16_t value) crstl_noexcept
 	{
-		return crstl_atomic_store16(target, value);
+		return (int16_t)crstl_atomic_store16(target, value);
 	}
 
 	inline int32_t atomic_store(int32_t volatile* target, int32_t value) crstl_noexcept
 	{
-		return crstl_atomic_store32(target, value);
+		return (int32_t)crstl_atomic_store32(target, value);
 	}
 
 	inline int64_t atomic_store(int64_t volatile* target, int64_t value) crstl_noexcept
 	{
-		return crstl_atomic_store64(target, value);
+		return (int64_t)crstl_atomic_store64(target, value);
 	}
 
 	// atomic_add: add value to variable pointed by target, and return the previous value of target
 
 	inline int8_t atomic_add(int8_t volatile* target, int8_t value) crstl_noexcept
 	{
-		return crstl_atomic_add8(target, value);
+		return (int8_t)crstl_atomic_add8(target, value);
 	}
 
 	inline int16_t atomic_add(int16_t volatile* target, int16_t value) crstl_noexcept
 	{
-		return crstl_atomic_add16(target, value);
+		return (int16_t)crstl_atomic_add16(target, value);
 	}
 
 	inline int32_t atomic_add(int32_t volatile* target, int32_t value) crstl_noexcept
 	{
-		return crstl_atomic_add32(target, value);
+		return (int32_t)crstl_atomic_add32(target, value);
 	}
 
 	inline int64_t atomic_add(int64_t volatile* target, int64_t value) crstl_noexcept
 	{
-		return crstl_atomic_add64(target, value);
+		return (int64_t)crstl_atomic_add64(target, value);
 	}
 
 	// atomic_sub: subtract value from variable pointed by target, and return the previous value of target
 
 	inline int8_t atomic_sub(int8_t volatile* target, int8_t value) crstl_noexcept
 	{
-		return crstl_atomic_sub8(target, value);
+		return (int8_t)crstl_atomic_sub8(target, value);
 	}
 
 	inline int16_t atomic_sub(int16_t volatile* target, int16_t value) crstl_noexcept
 	{
-		return crstl_atomic_sub16(target, value);
+		return (int16_t)crstl_atomic_sub16(target, value);
 	}
 
 	inline int32_t atomic_sub(int32_t volatile* target, int32_t value) crstl_noexcept
 	{
-		return crstl_atomic_sub32(target, value);
+		return (int32_t)crstl_atomic_sub32(target, value);
 	}
 
 	inline int64_t atomic_sub(int64_t volatile* target, int64_t value) crstl_noexcept
 	{
-		return crstl_atomic_sub64(target, value);
+		return (int64_t)crstl_atomic_sub64(target, value);
 	}
 
 	// atomic_and: AND value with variable pointed by target, and return the previous value of target
 
 	inline int8_t atomic_and(int8_t volatile* target, int8_t value) crstl_noexcept
 	{
-		return crstl_atomic_and8(target, value);
+		return (int8_t)crstl_atomic_and8(target, value);
 	}
 
 	inline int16_t atomic_and(int16_t volatile* target, int16_t value) crstl_noexcept
 	{
-		return crstl_atomic_and16(target, value);
+		return (int16_t)crstl_atomic_and16(target, value);
 	}
 
 	inline int32_t atomic_and(int32_t volatile* target, int32_t value) crstl_noexcept
 	{
-		return crstl_atomic_and32(target, value);
+		return (int32_t)crstl_atomic_and32(target, value);
 	}
 
 	inline int64_t atomic_and(int64_t volatile* target, int64_t value) crstl_noexcept
 	{
-		return crstl_atomic_and64(target, value);
+		return (int64_t)crstl_atomic_and64(target, value);
 	}
 
 	// atomic_or: OR value with variable pointed by target, and return the previous value of target
 
 	inline int8_t atomic_or(int8_t volatile* target, int8_t value) crstl_noexcept
 	{
-		return crstl_atomic_or8(target, value);
+		return (int8_t)crstl_atomic_or8(target, value);
 	}
 
 	inline int16_t atomic_or(int16_t volatile* target, int16_t value) crstl_noexcept
 	{
-		return crstl_atomic_or16(target, value);
+		return (int16_t)crstl_atomic_or16(target, value);
 	}
 
 	inline int32_t atomic_or(int32_t volatile* target, int32_t value) crstl_noexcept
 	{
-		return crstl_atomic_or32(target, value);
+		return (int32_t)crstl_atomic_or32(target, value);
 	}
 
 	inline int64_t atomic_or(int64_t volatile* target, int64_t value) crstl_noexcept
 	{
-		return crstl_atomic_or64(target, value);
+		return (int64_t)crstl_atomic_or64(target, value);
 	}
 
 	// atomic_xor: XOR value with variable pointed by target, and return the previous value of target
 
 	inline int8_t atomic_xor(int8_t volatile* target, int8_t value) crstl_noexcept
 	{
-		return crstl_atomic_xor8(target, value);
+		return (int8_t)crstl_atomic_xor8(target, value);
 	}
 
 	inline int16_t atomic_xor(int16_t volatile* target, int16_t value) crstl_noexcept
 	{
-		return crstl_atomic_xor16(target, value);
+		return (int16_t)crstl_atomic_xor16(target, value);
 	}
 
 	inline int32_t atomic_xor(int32_t volatile* target, int32_t value) crstl_noexcept
 	{
-		return crstl_atomic_xor32(target, value);
+		return (int32_t)crstl_atomic_xor32(target, value);
 	}
 
 	inline int64_t atomic_xor(int64_t volatile* target, int64_t value) crstl_noexcept
 	{
-		return crstl_atomic_xor64(target, value);
+		return (int64_t)crstl_atomic_xor64(target, value);
 	}
 
 	template<size_t N> struct atomic_type {};
